move epollfd implementation details into epollfd_ctx.c
