+++ src/epollfd_ctx.h
@@ -1,16 +1,39 @@
 #ifndef EPOLLFD_CTX_H_
 #define EPOLLFD_CTX_H_
 
+#define SHIM_SYS_SHIM_HELPERS
+#include <sys/epoll.h>
+
+#include <sys/tree.h>
+
 #include <errno.h>
+#include <stdint.h>
 
 #include <pthread.h>
+
+struct registered_fds_node_;
+typedef struct registered_fds_node_ RegisteredFDsNode;
+
+struct registered_fds_node_ {
+	RB_ENTRY(registered_fds_node_) entry;
+	int fd;
+	uint16_t flags;
+};
+
+typedef RB_HEAD(registered_fds_set_, registered_fds_node_) RegisteredFDsSet;
 
 typedef struct {
 	int kq; // non owning
 	pthread_mutex_t mutex;
+	RegisteredFDsSet registered_fds;
 } EpollFDCtx;
 
-errno_t epollfd_ctx_init(EpollFDCtx *signalfd, int kq);
-errno_t epollfd_ctx_terminate(EpollFDCtx *signalfd);
+errno_t epollfd_ctx_init(EpollFDCtx *epollfd, int kq);
+errno_t epollfd_ctx_terminate(EpollFDCtx *epollfd);
+
+errno_t epollfd_ctx_ctl(EpollFDCtx *epollfd, int op, int fd2,
+    struct epoll_event *ev);
+errno_t epollfd_ctx_wait(EpollFDCtx *epollfd, struct epoll_event *ev, int cnt,
+    int *actual_cnt);
 
 #endif
